+++ memorybound/memory_kernel.h
@@ -1,3 +1,3 @@
 #define MS 500
 
-extern "C" void matrixMulti_caller(float *Ma, float *Mb, float *Mc);
+extern "C" void matrixMulti_caller(double *Ma, double *Mb, double *Mc);
